feat: 🚀 1단계 - DB 연동하기
