+++ src/main/resources/application.properties
@@ -1,0 +1,4 @@
+spring.h2.console.enabled=true
+spring.datasource.url=jdbc:h2:mem:testdb;MODE=MySQL
+spring.datasource.driver-class-name=org.h2.Driver
+
